Support RFC4361 and RFC6842 on the server.
This change involves using the clientid, if provided by the client, as the
sole identifier for the owner of a given lease; a database format change
is required, as the previous unique key was the mac address.

A new ClientID class is introduced that keeps track of whether the client
provided a clientid, the client mac address, and the effective clientid
(a clientid is generated from the mac address if none exists in order
to support legacy clients).

The Lua scripting API changes a bit as well.  The main callback functions
grow an extra argument that allow both the mac and the effective clientid
that serves as the lease database primary key to be passed as arguments.

The server follows RFC6842 and will attach a clientid option to replies
destined to clients that themselves set the clientid option.  Clients
that set no clientid will not find themselves receiving responses that
have a clientid option.

All of these changes require quite a lot of code shuffling.