Support DHCPv6 and IPv6 router advertisements in addition to DHCPv4.
This was done by extending nrad6 so that it supports DHCPv4.

That history can be seen in the nrad6 repository, which will now
effectively be dead.

ndhs is no longer programmable in Lua and no longer stores dynamic
leases in a SQLite database.  Instead it uses a custom declarative
configuration file format and stores dynamic leases in an easily
parsed flat text file.

The old version of ndhs still exists in a 'legacy' branch.