nlsocket: get_interface_addresses() is a better name
add_interface() is very confusing; nlsocket has already gathered
the list of basic interfaces and associated data.  All that is
being requested at this point are the list of ip addresses
for the given interface.

While we're at it, it saves a table traversal to use ifindex
rather than ifname here.