dhcp_state, dhcp6: Option NTP_SERVER (56) isn't supported.
There was never a way to specify this option in the config file,
so remove support for it.

It's asymmetric to how dhcp4 works.  If it catches on, then it
can be added back.  RFC5908 does deprecate the alternative
SNTP_SERVERS (31) option, but it still is widely supported
and despite the name, does specify NTP servers, not SNTP.

Further, remove the use of std::vector<std::string> in
dhcp_state.cpp.  Instead, intrusive single-linked lists of
null-terminated strings are used.  These have the same
number of indirections (ignoring SSO) and are only used
until the search blobs are constructed at init time, anyway.