cfg: Correct a regression and an old latent bug.
Latent bug:
The DUID format with dashes (as is pretty-printed on win32) was
accepted by the parser, but was not properly stripped of dashes.

Fix by rejecting dashes; it's simple enough to just remove them
with s/-//g or the equivalent if copy-pasting.

Regression:
Parsing of v6 static leases was wrong and ignored the proper
iaid value.  This happened in the recent conversions.