dynlease: Fix a crash that can happen when gcing old dynleases.
We can deref a null pointer in these cases since p can be assigned
to null within the body of the loop.

Introduced with the rewrite to C.  There don't appear to be any
other similar bugs, and this crash isn't exploitable as there
are no inputs for an attacker to control and it is always a
null-pointer-deref that causes the segfault.