Don't busy-spin if s6_notify is spuriously set.
This one was tricky to figure out.  If s6_notify is set but a
notification file descriptor was not provided to the ndhs process,
then it is very likely that that fd number will be taken by
one of the listeners.  In that case, ndhs will end up writing a
s6 notification to that listener fd and then it will close that
listener fd.  Afterwards, the main poll() will end up busy-spinning.

Arguably this isn't a bug since the user provided a bad configuration
file.  But in the interests of better QoI and user experience,
detect this misconfiguration by comparing the s6 notify fd value to
listeners.  If it matches a listener fd, don't try to perform
a s6 notification.